[wip] step 2 QA passed, can create 2 tg in the selected regions
